fix - refactored Join Types code to break it into smaller components
